fix(location): refactor location to request location update & safety nullable
