+++ app/src/main/AndroidManifest.xml
@@ -17,9 +17,6 @@
         android:theme="@style/Theme.KotlinCorePlatform"
         tools:targetApi="31">
         <activity
-            android:name="com.fadlurahmanfdev.example.presentation.LocationActivity"
-            android:exported="false" />
-        <activity
             android:name="com.fadlurahmanfdev.example.presentation.MainActivity"
             android:exported="true">
             <intent-filter>
